lib/cbfs: more cleanup for 32/64 issues
Reviewed-on: http://review.coreboot.org/7068
Tested-by: build bot (Jenkins)

+++ src/lib/cbfs.c
@@ -123,7 +123,7 @@ void * cbfs_load_stage(struct cbfs_media *media, const char *name)
 		cbfs_get_file_content(media, name, CBFS_TYPE_STAGE, NULL);
 	/* this is a mess. There is no ntohll. */
 	/* for now, assume compatible byte order until we solve this. */
-	uint32_t entry;
+	uintptr_t entry;
 	uint32_t final_size;
 
 	if (stage == NULL)
@@ -137,7 +137,7 @@ void * cbfs_load_stage(struct cbfs_media *media, const char *name)
 	final_size = cbfs_decompress(stage->compression,
 				     ((unsigned char *) stage) +
 				     sizeof(struct cbfs_stage),
-				     (void *) (uint32_t) stage->load,
+				     (void *) (uintptr_t) stage->load,
 				     stage->len);
 	if (!final_size)
 		return (void *) -1;
